requirements.txt
# Specify Python package requirements for your project here (e.g., Mako==1.1.1). If your project doesn't require these, you can leave this file unchanged or delete it.
langchain>=0.3.25,<1.0.0
langchain-openai>=0.0.5,<1.0.0
langchain-chroma>=0.2.4,<1.0.0
langchain_community>=0.3.26,<1.0.0
streamlit>=1.31.0
pdfplumber>=0.11.6
chromadb>=1.0.10
ruff==0.11.11
python-dotenv>=1.0.0